requirements.txt
flask
gunicorn
psycopg2-binary
pdfplumber
python-docx
joblib
scikit-learn
pandas
numpy
